+++ packages/solid-snowfall/src/hooks.ts
@@ -1,9 +1,9 @@
-import { Accessor, createEffect, createMemo, createSignal, JSX, on, onCleanup } from "solid-js";
+import { Accessor, createEffect, createSignal, on, onCleanup } from "solid-js";
+import { Store } from "solid-js/store";
 
-import isEqual from "fast-deep-equal";
+import { trackStore } from "@solid-primitives/deep";
 
-import { snowfallBaseStyle } from "./config";
-import Snowflake, { SnowflakeConfig } from "./Snowflake";
+import { Snowflake, SnowflakeConfig } from "./Snowflake";
 import { getSize } from "./utils";
 
 /**
@@ -33,7 +33,7 @@ const makeSnowflakesArray = (canvasRef: HTMLCanvasElement, amount: number, confi
 export const createSnowFlakes = (
   canvasRef: Accessor<HTMLCanvasElement>,
   amount: Accessor<number>,
-  config: Accessor<SnowflakeConfig>,
+  config: Store<SnowflakeConfig>,
 ) => {
   const [snowflakes, setSnowflakes] = createSignal<Snowflake[]>([]);
 
@@ -44,7 +44,7 @@ export const createSnowFlakes = (
         const sizeDifference = amount() - snowflakes.length;
 
         if (sizeDifference > 0) {
-          return [...snowflakes, ...makeSnowflakesArray(canvasRef(), sizeDifference, config())];
+          return [...snowflakes, ...makeSnowflakesArray(canvasRef(), sizeDifference, config)];
         }
 
         if (sizeDifference < 0) {
@@ -58,9 +58,10 @@ export const createSnowFlakes = (
 
   // Handle change of config
   createEffect(() => {
+    trackStore(config);
     setSnowflakes((snowflakes) =>
       snowflakes.map((snowflake) => {
-        snowflake.updateConfig(config());
+        snowflake.updateConfig(config);
         return snowflake;
       }),
     );
@@ -103,38 +104,3 @@ export const createComponentSize = (ref: Accessor<HTMLElement>) => {
 
   return size;
 };
-
-/**
- * Returns the height and width of a HTML element, uses the `ResizeObserver` api if available to detect changes to the
- * size. Falls back to listening for resize events on the window.
- * @param ref A ref to the HTML element to be measured
- */
-export const createSnowFallStyle = (
-  overrides?: Accessor<JSX.CSSProperties | undefined>,
-): Accessor<JSX.CSSProperties> => {
-  const styles = createMemo(() => ({
-    ...snowfallBaseStyle,
-    ...(overrides?.() || {}),
-  }));
-
-  return styles;
-};
-
-/**
- * Utility hook to stabilize a reference to a value, the returned value will always match the input value
- * but (unlike an inline object) will maintain [SameValueZero](http://ecma-international.org/ecma-262/7.0/#sec-samevaluezero)
- * equality until a change is made.
- *
- * @example
- *
- * const obj = useDeepMemo({ foo: 'bar', bar: 'baz' }) // <- inline object creation
- * const prevValue = usePrevious(obj) // <- value from the previous render
- * console.log(obj === prevValue) // <- always logs true until value changes
- */
-export function createDeepCompareMemo<T>(value: Accessor<T>): Accessor<T> {
-  const result = createMemo(() => value(), undefined, {
-    equals: (a, b) => isEqual(a, b),
-  });
-
-  return result as Accessor<T>;
-}