+++ packages/solid-snowfall/vitest.config.ts
@@ -1,0 +1,15 @@
+/// <reference types="vitest" />
+
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    environment: "jsdom",
+    globals: true,
+    testTransformMode: { web: ["/.[jt]sx?$/"] },
+    // if you have few tests, try commenting one
+    // or both out to improve performance:
+    // threads: false,
+    // isolate: false,
+  },
+});